file_handler.py
import os
from pdfminer.high_level import extract_text as pdf_extract_text
import openai  # For image analysis if vision is added

def handle_file_upload(file_path):
    """
    Handle file uploads: parse text files, PDFs, and placeholder for images.
    Returns extracted text or description.
    """
    if not os.path.exists(file_path):
        return "File not found."

    ext = os.path.splitext(file_path)[1].lower()
    if ext == '.txt':
        with open(file_path, 'r', encoding='utf-8') as f:
            return f.read()
    elif ext == '.pdf':
        try:
            return pdf_extract_text(file_path)
        except Exception as e:
            return f"Error parsing PDF: {e}"
    elif ext in ['.jpg', '.jpeg', '.png']:
        # Placeholder for image analysis
        return "Image uploaded. Vision analysis not yet implemented. Describe the image manually."
        # Future: Use OpenAI Vision API
        # client = openai.OpenAI()
        # with open(file_path, "rb") as image_file:
        #     response = client.chat.completions.create(
        #         model="gpt-4-vision-preview",
        #         messages=[{"role": "user", "content": [{"type": "image_url", "image_url": {"url": f"data:image/jpeg;base64,{base64.b64encode(image_file.read()).decode()}"}}]}]
        #     )
        #     return response.choices[0].message.content
    else:
        return "Unsupported file type. Supported: .txt, .pdf, .jpg, .png"
