requirements.txt
google-api-python-client==2.110.0
python-dotenv==1.0.0
openai==1.12.0
pyttsx3==2.90
groq==0.4.1
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
pdfminer.six==20231228
sentence-transformers
faiss-cpu==1.9.0.post1
googletrans==4.0.0rc1
rich==13.7.0
newsapi-python==0.2.7
yfinance==0.2.28
wikipedia-api==0.6.0
selenium==4.15.2
webdriver-manager==4.0.1
